feat/refactor/fix: make it worked on Windows
* fix: get_time overflow
