+++ Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "dnspeep"
-version = "0.1.2"
+version = "0.1.3"
 authors = ["Julia Evans <julia@jvns.ca>"]
 edition = "2018"
 
 [dependencies]
-pcap = { git = "https://github.com/jvns/pcap", features=["capture-stream"] }
+pcap = { git = "https://github.com/jvns/pcap"}
 libc = "0.2.80"
 dns-message-parser = "~0.5"
 etherparse = "0.9.0"
@@ -16,3 +16,8 @@ getopts = "0.2.17"
 bytes = "~1.0"
 hex = "~0.4"
 chrono = "0.4.19"
+
+[features]
+default = []
+pcap_windows = []
+pcap_unix = ["pcap/capture-stream"]
